Move dependencies into galaxy.webapps.galaxy.api
